archive/duplicates/20251107/extras/requirements-1.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
pydantic==2.9.2
python-dotenv==1.0.1
pytest==8.3.3
PyJWT==2.9.0
